Committing svn revision 32914: Revert bad [32913] for ZCP-9615

+++ common/IStreamAdapter.cpp
@@ -49,8 +49,6 @@
 
 #include "platform.h"
 
-#include <algorithm>
-
 #include <mapicode.h>
 #include <mapidefs.h>
 #include <mapiguid.h>
@@ -77,7 +75,7 @@ ULONG IStreamAdapter::Release() { return 1; }
 
 HRESULT IStreamAdapter::Read(void *pv, ULONG cb, ULONG *pcbRead)
 {
-	size_t toread = std::min(cb, (ULONG)(m_str.size() - m_pos));
+	size_t toread = std::min(cb, m_str.size() - m_pos);
 	
 	memcpy(pv, m_str.data() + m_pos, toread);
